refactor: move is inner link logic out of installMultiple
