+++ src/api/install.ts
@@ -4,6 +4,8 @@ import logger from 'pnpm-logger'
 import pLimit = require('p-limit')
 import npa = require('npm-package-arg')
 import symlinkDir from 'symlink-dir'
+import pFilter = require('p-filter')
+import getLinkTarget = require('get-link-target')
 import {PnpmOptions, StrictPnpmOptions, Dependencies} from '../types'
 import createGot from '../network/got'
 import getContext, {PnpmContext} from './getContext'
@@ -147,11 +149,11 @@ async function installInContext (
     resolvedDependencies,
     offline: opts.offline,
   }
+  const nonLinkedPkgs = await pFilter(packagesToInstall, (spec: PackageSpec) => !spec.name || isInnerLink(nodeModulesPath, spec.name))
   const pkgs: InstalledPackage[] = await installMultiple(
     installCtx,
-    packagesToInstall,
+    nonLinkedPkgs,
     optionalDependencies,
-    nodeModulesPath,
     installOpts
   )
   await linkPackages(pkgs, nodeModulesPath, installCtx.installs, {
@@ -235,6 +237,23 @@ async function installInContext (
       npmRun('prepublish', ctx.root)
     }
   }
+}
+
+async function isInnerLink (modules: string, depName: string) {
+  let linkTarget: string
+  try {
+    const linkPath = path.join(modules, depName)
+    linkTarget = await getLinkTarget(linkPath)
+  } catch (err) {
+    if (err.code === 'ENOENT') return true
+    throw err
+  }
+
+  if (linkTarget.startsWith(modules)) {
+    return true
+  }
+  logger.info(`${depName} is linked to ${modules} from ${linkTarget}`)
+  return false
 }
 
 async function createInstallCmd (opts: StrictPnpmOptions, shrinkwrap: Shrinkwrap): Promise<InstallContext> {
